refactor code to follow MVP structure
