+++ app/src/main/java/com/sdsmdg/kd/trianglifyexample/MainActivity.java
@@ -3,13 +3,11 @@ package com.sdsmdg.kd.trianglifyexample;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
-import android.view.View;
 import android.widget.CheckBox;
-import android.widget.Checkable;
 import android.widget.CompoundButton;
 import android.widget.SeekBar;
 
-import com.sdsmdg.kd.trianglify.TrianglifyView;
+import com.sdsmdg.kd.trianglify.views.TrianglifyView;
 import com.sdsmdg.kd.trianglify.models.Palette;
 
 public class MainActivity extends AppCompatActivity {
@@ -17,7 +15,7 @@ public class MainActivity extends AppCompatActivity {
             Palette.PuBu, Palette.BuPu, Palette.RdPu, Palette.PuRd, Palette.OrRd, Palette.YlOrRd,
             Palette.YlOrBr, Palette.Purples, Palette.Blues, Palette.Greens, Palette.Oranges,
             Palette.Reds, Palette.Greys, Palette.PuOr, Palette.BrBG, Palette.PRGn, Palette.PiYG,
-            Palette.RdBu, Palette.RdGy, Palette.RdYlBu, Palette.Spectral, Palette.RdYlGn};
+            Palette.RdBu, Palette.RdGy, Palette.RdYlBu, Palette.Spectral, Palette.RdYlGn, Palette.Yl};
 
     public TrianglifyView trianglifyView;
     private SeekBar varianceSeekBar;
@@ -114,7 +112,7 @@ public class MainActivity extends AppCompatActivity {
         fillCheckBox.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
             @Override
             public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                trianglifyView.setFillStroke(isChecked);
+                trianglifyView.setFillTriangle(isChecked);
                 trianglifyView.invalidate();
             }
         });
