+++ app/src/main/res/layout/activity_main.xml
@@ -61,7 +61,7 @@
         android:text="Fill Triangles"
         />
 
-    <com.sdsmdg.kd.trianglify.TrianglifyView
+    <com.sdsmdg.kd.trianglify.views.TrianglifyView
         android:id="@+id/trianglify_main_view"
         android:layout_marginTop="20dp"
         android:layout_centerHorizontal="true"
