--- trianglify/src/main/java/com/sdsmdg/kd/trianglify/generators/Triangulator.java
@@ -1,16 +1,0 @@
-package com.sdsmdg.kd.trianglify.generators;
-
-import com.sdsmdg.kd.trianglify.models.Grid;
-import com.sdsmdg.kd.trianglify.models.Triangulation;
-
-/**
- * Created by shyam on 12-Mar-17.
- */
-
-public interface Triangulator  {
-
-    public Triangulator setGrid(Grid grid);
-
-    public Triangulation getTriangulation();
-
-}
