+++ trianglify/src/main/java/com/sdsmdg/kd/trianglify/utilities/colorizers/ExtendedColor.java
@@ -1,6 +1,4 @@
-package com.sdsmdg.kd.trianglify.utilities;
-
-import android.graphics.Color;
+package com.sdsmdg.kd.trianglify.utilities.colorizers;
 
 /**
  * <h1>Title</h1>
