+++ trianglify/src/main/java/com/sdsmdg/kd/trianglify/utilities/colorizers/Point.java
@@ -1,4 +1,4 @@
-package com.sdsmdg.kd.trianglify.utilities;
+package com.sdsmdg.kd.trianglify.utilities.colorizers;
 
 /**
  * <h1>Title</h1>
