+++ trianglify/src/main/java/com/sdsmdg/kd/trianglify/utilities/patterns/Patterns.java
@@ -1,0 +1,16 @@
+package com.sdsmdg.kd.trianglify.utilities.patterns;
+
+
+
+import com.sdsmdg.kd.trianglify.utilities.triangulator.Vector2D;
+
+import java.util.List;
+
+/**
+ * Created by suyash on 17/3/17.
+ */
+
+
+public interface Patterns {
+    public List<Vector2D> generate();
+}
