+++ trianglify/src/main/java/com/sdsmdg/kd/trianglify/utilities/triangulator/NotEnoughPointsException.java
@@ -1,4 +1,4 @@
-package com.sdsmdg.kd.trianglify.models.triangulator;
+package com.sdsmdg.kd.trianglify.utilities.triangulator;
 
 /**
  * Exception thrown by the Delaunay triangulator when it is initialized with
