+++ trianglify/src/main/java/com/sdsmdg/kd/trianglify/views/TrianglifyView.java
@@ -1,32 +1,18 @@
-package com.sdsmdg.kd.trianglify;
+package com.sdsmdg.kd.trianglify.views;
 
 import android.content.Context;
 import android.content.res.TypedArray;
 import android.graphics.Canvas;
-import android.graphics.Color;
 import android.graphics.Paint;
 import android.graphics.Path;
 import android.util.AttributeSet;
-import android.util.Log;
 import android.view.View;
-import android.widget.Toast;
 
-import com.sdsmdg.kd.trianglify.models.Grid;
+import com.sdsmdg.kd.trianglify.presenters.Presenter;
+import com.sdsmdg.kd.trianglify.R;
 import com.sdsmdg.kd.trianglify.models.Palette;
 import com.sdsmdg.kd.trianglify.models.Triangulation;
-import com.sdsmdg.kd.trianglify.models.triangulator.DelaunayTriangulator;
-import com.sdsmdg.kd.trianglify.models.triangulator.NotEnoughPointsException;
-import com.sdsmdg.kd.trianglify.models.triangulator.Triangle2D;
-import com.sdsmdg.kd.trianglify.models.triangulator.Vector2D;
-import com.sdsmdg.kd.trianglify.patterns.Patterns;
-import com.sdsmdg.kd.trianglify.patterns.Rectangle;
-import com.sdsmdg.kd.trianglify.utilities.colorizers.Colorizer;
-import com.sdsmdg.kd.trianglify.utilities.colorizers.FixedPointsColorizer;
-
-import java.util.List;
-import java.util.Vector;
-
-import static android.content.ContentValues.TAG;
+import com.sdsmdg.kd.trianglify.utilities.triangulator.Triangle2D;
 
 public class TrianglifyView extends View implements TrianglifyViewInterface{
     int bleedX;
@@ -46,7 +32,6 @@ public class TrianglifyView extends View implements TrianglifyViewInterface{
             Palette.RdBu, Palette.RdGy, Palette.RdYlBu, Palette.Spectral, Palette.RdYlGn};
     Palette palette;
     Triangulation triangulation;
-
     Presenter presenter;
 
     public TrianglifyView(Context context, AttributeSet attrs) {
@@ -56,20 +41,12 @@ public class TrianglifyView extends View implements TrianglifyViewInterface{
         this.presenter = new Presenter(this);
     }
 
-    public Palette getPalette() {
-        return palette;
-    }
-
-    public void setPalette(Palette palette) {
-        this.palette = palette;
-    }
-
     @Override
-    protected void onSizeChanged(int w, int h, int oldw, int oldh) {
-        super.onSizeChanged(w, h, oldw, oldh);
+    protected void onSizeChanged(int width, int height, int oldWidth, int oldHeight) {
+        super.onSizeChanged(width, height, oldWidth, oldHeight);
 
-        gridWidth = w;
-        gridHeight =h;
+        gridWidth = width;
+        gridHeight =height;
     }
 
     public void attributeSetter(TypedArray typedArray){
@@ -88,7 +65,14 @@ public class TrianglifyView extends View implements TrianglifyViewInterface{
         }
     }
 
+    @Override
+    public Palette getPalette() {
+        return palette;
+    }
 
+    public void setPalette(Palette palette) {
+        this.palette = palette;
+    }
 
     @Override
     public int getCellSize() {
@@ -153,6 +137,7 @@ public class TrianglifyView extends View implements TrianglifyViewInterface{
         this.variance = variance;
     }
 
+    @Override
     public int getPaletteNumber() {
         return paletteNumber;
     }
@@ -195,13 +180,19 @@ public class TrianglifyView extends View implements TrianglifyViewInterface{
         generateAndPlot(canvas);
     }
 
+    void generateAndPlot(Canvas canvas) {
+        generate();
+        plotOnCanvas(canvas);
+    }
+
     void generate() {
         this.triangulation = presenter.getSoup();
     }
 
-    void generateAndPlot(Canvas canvas) {
-        generate();
-        plotOnCanvas(canvas);
+    void plotOnCanvas(Canvas canvas) {
+        for (int i = 0; i < triangulation.getTriangleList().size(); i++){
+            drawTriangle(canvas, triangulation.getTriangleList().get(i));
+        }
     }
 
     public void drawTriangle(Canvas canvas, Triangle2D triangle2D) {
@@ -235,12 +226,5 @@ public class TrianglifyView extends View implements TrianglifyViewInterface{
         path.close();
 
         canvas.drawPath(path, paint);
-    }
-
-    void plotOnCanvas(Canvas canvas) {
-        for (int i = 0; i < triangulation.getTriangleList().size(); i++){
-            drawTriangle(canvas, triangulation.getTriangleList().get(i));
-        }
-
     }
 }
