+++ trianglify/src/main/java/com/sdsmdg/kd/trianglify/views/TrianglifyViewInterface.java
@@ -1,4 +1,4 @@
-package com.sdsmdg.kd.trianglify;
+package com.sdsmdg.kd.trianglify.views;
 
 /**
  * Created by suyash on 18/3/17.
@@ -6,14 +6,10 @@ package com.sdsmdg.kd.trianglify;
 
 import com.sdsmdg.kd.trianglify.models.Palette;
 import com.sdsmdg.kd.trianglify.models.Triangulation;
-import com.sdsmdg.kd.trianglify.patterns.Patterns;
 
-interface TrianglifyViewInterface {
+public interface TrianglifyViewInterface {
     int GRID_RECTANGLE = 0;
     int GRID_CIRCLE = 1;
-
-    int TRIANGULATION_DELAUNAY = 0;
-    int TRIANGULATION_SHAMOS = 0;
 
     int getBleedX();
     int getBleedY();
